// asr.webapp.v2/src/app/modules/inventory/reports/goods/goods-inventory.types.ts
import { Paging } from "app/core/type/paging/paging.type";
import { BaseModel } from "app/core/type/base/BaseModel";

export class GoodsInventory extends BaseModel {
    reportTypeId: number;
    reportTypeName: string;
    goodsId: number;
    goodsCode: string;
    technicalNumber: string;
    categoryGoodsCode: string;
    goodsName: string;
    categoryId: number;
    categoryName: string;
    warehouseIds: Array<number>;
    warehouseId: number;
    warehouseName: string;
    initialStock: number;
    stock: number;
    stockChanges: number;
    reserve: number;
    preReserve: number;
    remained: number;
    initialUnitPrice: number;
    unitPrice: number;
    totalPrice: number;
    currency: string;
    status: number;
    page: Paging;
    coverIds: Array<number>;
    stockChangesIdList: Array<number>;
    coverId: number;
    coverName: string;
    unitName: string;
    initialUnitPriceFrom: number;
    initialUnitPriceTo: number;
    finalUnitPriceFrom: number;
    finalUnitPriceTo: number;
    initialStockFrom: number;
    initialStockTo: number;
    finalStockFrom: number;
    finalStockTo: number;
    stockChangesFrom: number;
    stockChangesTo: number;
    totalPriceFrom: number;
    totalPriceTo: number;

}
export class GoodsInventoryHistory extends BaseModel {
    logId: number;
    warehouseInventoryId: number;
    goodsId: number;
    goodsName: string;
    goodsCode: string;
    technicalNumber: string;
    coverId: number;
    coverName: string;
    warehouseId: number;
    warehouseName: string;
    stock: number;
    reserve: number;
    preReserve: number;
    ownerId: number;
    ownerName: string;
    ownerTypeId: number;
    ownerTypeName: string;
    unitId: number;
    unitName: string;
    initialStock: number;
    initialPrice: number;
    requestTypeId: number;
    requestTypeName: string;
    warehouseIds: Array<number>;
    coverIds: Array<number>;
    initialPriceFrom: number;
    initialPriceTo: number;
    initialStockFrom: number;
    initialStockTo: number;
    finalStockFrom: number;
    finalStockTo: number;
    page: Paging;
}